requirements.txt
requests
bs4
lxml
selenium
pandas
html5lib
openpyxl
python-dotenv===0.20.0
xlsxwriter